NM added options.

+++ sen2_geocor.py
@@ -9,6 +9,11 @@ except Exception:
 from subprocess import check_output,call
 from optparse import OptionParser,IndentedHelpFormatter
 
+# Default values
+#REF_DATA_MIN = None
+REF_DATA_MIN = 0.1 # for WorldView DN image
+RESAMPLING = 'cubic'
+
 # Read options
 parser = OptionParser(formatter=IndentedHelpFormatter(max_help_position=200,width=200))
 parser.set_usage('Usage: %prog reference_geotiff_file target_geotiff_file [options]')
@@ -26,8 +31,17 @@ parser.add_option('--shift_width',default=None,type='int',help='Max shift width 
 parser.add_option('--shift_height',default=None,type='int',help='Max shift height in target pixel (%default)')
 parser.add_option('--margin_width',default=None,type='int',help='Margin width in target pixel (%default)')
 parser.add_option('--margin_height',default=None,type='int',help='Margin height in target pixel (%default)')
+parser.add_option('--ref_data_min',default=REF_DATA_MIN,type='float',help='Minimum reference data value (%default)')
+parser.add_option('--ref_data_max',default=None,type='float',help='Maximum reference data value (%default)')
+parser.add_option('-g','--use_gcps',default=None,help='GCP file name to use (%default)')
+parser.add_option('-G','--save_gcps',default=None,help='GCP file name to save (%default)')
+parser.add_option('-e','--trg_epsg',default=None,help='Target EPSG (guessed from target data)')
 parser.add_option('-r','--rthr',default=None,type='float',help='Threshold of correlation coefficient (%default)')
 parser.add_option('-E','--feps',default=None,type='float',help='Step length for curve_fit (%default)')
+parser.add_option('-n','--npoly',default=None,type='int',help='Order of polynomial used for warping between 1 and 3 (selected based on the number of GCPs)')
+parser.add_option('-R','--resampling',default=RESAMPLING,help='Resampling method (%default)')
+parser.add_option('--refine_gcps',default=None,type='float',help='Tolerance to refine GCPs for polynomial interpolation (%default)')
+parser.add_option('--tps',default=False,action='store_true',help='Use thin plate spline transformer (%default)')
 parser.add_option('-d','--debug',default=False,action='store_true',help='Debug mode (%default)')
 (opts,args) = parser.parse_args()
 if len(args) < 1:
@@ -36,49 +50,63 @@ if len(args) < 1:
 ref_fnam = args[0]
 trg_fnam = args[1]
 tmp_fnam = 'tmp.tif'
+if opts.trg_epsg is None:
+    ds = gdal.Open(trg_fnam)
+    prj = ds.GetProjection()
+    srs = osr.SpatialReference(wkt=prj)
+    opts.trg_epsg = srs.GetAttrValue('AUTHORITY',1)
+    ds = None # close dataset
+
 out_fnam = os.path.splitext(os.path.basename(trg_fnam))[0]+'_geocor.tif'
+if opts.use_gcps is not None:
+    fnam = opts.use_gcps
+else:
+    command = 'find_gcps.py'
+    command += ' '+ref_fnam
+    command += ' '+trg_fnam
+    command += ' -v'
+    if opts.ref_band is not None:
+        command += ' --ref_band {}'.format(opts.ref_band)
+    if opts.trg_band is not None:
+        command += ' --trg_band {}'.format(opts.trg_band)
+    if opts.trg_indx_start is not None:
+        command += ' --trg_indx_start {}'.format(opts.trg_indx_start)
+    if opts.trg_indx_stop is not None:
+        command += ' --trg_indx_stop {}'.format(opts.trg_indx_stop)
+    if opts.trg_indx_step is not None:
+        command += ' --trg_indx_step {}'.format(opts.trg_indx_step)
+    if opts.trg_indy_start is not None:
+        command += ' --trg_indy_start {}'.format(opts.trg_indy_start)
+    if opts.trg_indy_stop is not None:
+        command += ' --trg_indy_stop {}'.format(opts.trg_indy_stop)
+    if opts.trg_indy_step is not None:
+        command += ' --trg_indy_step {}'.format(opts.trg_indy_step)
+    if opts.subset_width is not None:
+        command += ' --subset_width {}'.format(opts.subset_width)
+    if opts.subset_height is not None:
+        command += ' --subset_height {}'.format(opts.subset_height)
+    if opts.shift_width is not None:
+        command += ' --shift_width {}'.format(opts.shift_width)
+    if opts.shift_height is not None:
+        command += ' --shift_height {}'.format(opts.shift_height)
+    if opts.margin_width is not None:
+        command += ' --margin_width {}'.format(opts.margin_width)
+    if opts.margin_height is not None:
+        command += ' --margin_height {}'.format(opts.margin_height)
+    if opts.ref_data_min is not None:
+        command += ' --ref_data_min {}'.format(opts.ref_data_min)
+    if opts.ref_data_max is not None:
+        command += ' --ref_data_max {}'.format(opts.ref_data_max)
+    if opts.rthr is not None:
+        command += ' --rthr {}'.format(opts.rthr)
+    if opts.feps is not None:
+        command += ' --feps {}'.format(opts.feps)
+    if opts.debug:
+        command += ' --debug'
+    out = check_output(command,shell=True)
+    fnam = StringIO(out.decode())
+xi,yi,xp,yp,dx,dy,r = np.loadtxt(fnam,unpack=True)
 
-command = 'sen2_geocor.py'
-command += ' '+ref_fnam
-command += ' '+trg_fnam
-command += ' -v'
-if opts.ref_band is not None:
-    command += ' --ref_band {}'.format(opts.ref_band)
-if opts.trg_band is not None:
-    command += ' --trg_band {}'.format(opts.trg_band)
-if opts.trg_indx_start is not None:
-    command += ' --trg_indx_start {}'.format(opts.trg_indx_start)
-if opts.trg_indx_stop is not None:
-    command += ' --trg_indx_stop {}'.format(opts.trg_indx_stop)
-if opts.trg_indx_step is not None:
-    command += ' --trg_indx_step {}'.format(opts.trg_indx_step)
-if opts.trg_indy_start is not None:
-    command += ' --trg_indy_start {}'.format(opts.trg_indy_start)
-if opts.trg_indy_stop is not None:
-    command += ' --trg_indy_stop {}'.format(opts.trg_indy_stop)
-if opts.trg_indy_step is not None:
-    command += ' --trg_indy_step {}'.format(opts.trg_indy_step)
-if opts.subset_width is not None:
-    command += ' --subset_width {}'.format(opts.subset_width)
-if opts.subset_height is not None:
-    command += ' --subset_height {}'.format(opts.subset_height)
-if opts.shift_width is not None:
-    command += ' --shift_width {}'.format(opts.shift_width)
-if opts.shift_height is not None:
-    command += ' --shift_height {}'.format(opts.shift_height)
-if opts.margin_width is not None:
-    command += ' --margin_width {}'.format(opts.margin_width)
-if opts.margin_height is not None:
-    command += ' --margin_height {}'.format(opts.margin_height)
-if opts.rthr is not None:
-    command += ' --rthr {}'.format(opts.rthr)
-if opts.feps is not None:
-    command += ' --feps {}'.format(opts.feps)
-if opts.debug:
-    command += ' --debug'
-out = check_output(command,shell=True)
-
-xi,yi,xp,yp,dx,dy,r = np.loadtxt(StringIO(out.decode()),unpack=True)
 command = 'gdal_translate'
 for i,j,x,y in zip(xi,yi,xp,yp):
     command += ' -gcp {} {} {} {}'.format(i,j,x,y)
@@ -87,13 +115,17 @@ command += ' '+tmp_fnam
 call(command,shell=True)
 
 command = 'gdalwarp'
-command += ' -t_srs EPSG:32748'
-command += ' -r cubic'
-command += ' -order 2'
-#command += ' -refine_gcps 0.5'
+command += ' -t_srs EPSG:{}'.format(opts.trg_epsg)
 command += ' -overwrite'
-#command += ' -tps'
+if opts.tps:
+    command += ' -tps'
+elif opts.npoly is not None:
+    command += ' -order {}'.format(opts.npoly)
+command += ' -r {}'.format(opts.resampling)
+if opts.refine_gcps is not None:
+    command += ' -refine_gcps {}'.format(opts.refine_gcps)
 command += ' '+tmp_fnam
 command += ' '+out_fnam
 call(command,shell=True)
-os.remove(tmp_fnam)
+if os.path.exists(tmp_fnam):
+    os.remove(tmp_fnam)
